Added option for showing states

+++ wp-content/plugins/state-city/state-city.php
@@ -40,6 +40,7 @@ class States_Cities extends WP_List_Table {
 		) );
 		
 		add_action('admin_menu', array(&$this, 'my_menu_pages'));
+
 	// Install plugin
 		register_activation_hook( __FILE__, array( $this, 'activate' ) );
 		register_deactivation_hook( __FILE__, array( $this, 'deactivate' ) );
@@ -67,11 +68,41 @@ class States_Cities extends WP_List_Table {
 			'cb'        => '<input type="checkbox" name="bulk[]" />', //Render a checkbox instead of text
 			'city'    => 'City',
 			'zip' => 'Zip',
-			'state'      => 'State Name',
-			'stateCode'      => 'State Code'
+			'state_name'      => 'State Name',
+			'state_code'      => 'State Code'
 		);
 		return $columns;
-	}	
+	}
+	public function get_columns_forState(){
+		$columns = array(
+			'cb'        => '<input type="checkbox" name="bulk[]" />', //Render a checkbox instead of text
+			'state_name'      => 'State Name',
+			'state_code'      => 'State Code',
+			'cities'      => 'Total Cities'
+		);
+		return $columns;
+	}
+	
+	function prepare_items_forState(){
+		global $wpdb;
+		$columns = $this->get_columns_forState();
+		$hidden = array();
+		$sortable = $this->get_sortable_columns_forState();
+		$this->process_bulk_action();
+		$this->_column_headers = array($columns, $hidden, $sortable);
+		$data = $this->listDataStates() ;
+		$this->items = $data;
+				
+		$per_page = $this->per_page ;	
+		$current_page = $this->get_pagenum();
+		$total_items = $this->getCountDataState();
+		$this->set_pagination_args( array(
+				'total_items' => $total_items,                  //WE have to calculate the total number of items
+				'per_page'    => $per_page,                     //WE have to determine how many items to show on a page
+				'total_pages' => ceil($total_items/$per_page)   //WE have to calculate the total number of pages
+		) );		
+	
+	}
 	public function prepare_items() {
 		global $wpdb;
 		
@@ -83,15 +114,8 @@ class States_Cities extends WP_List_Table {
 		$data = $this->listData() ;
 		$this->items = $data;
 	
-			//Explicitly written within function
-			function usort_reorder($a,$b){
-					$orderby = (!empty($_REQUEST['orderby'])) ? $_REQUEST['orderby'] : 'title'; //If no sort, default to title
-					$order = (!empty($_REQUEST['order'])) ? $_REQUEST['order'] : 'asc'; //If no order, default to asc
-					$result = strcmp($a[$orderby], $b[$orderby]); //Determine sort order
-					return ($order==='asc') ? $result : -$result; //Send final sort direction to usort
-			}
 
-			usort($data, 'usort_reorder');		
+			
 		$per_page = $this->per_page ;	
 		$current_page = $this->get_pagenum();
 		$total_items = $this->getCountData();
@@ -100,6 +124,14 @@ class States_Cities extends WP_List_Table {
 				'per_page'    => $per_page,                     //WE have to determine how many items to show on a page
 				'total_pages' => ceil($total_items/$per_page)   //WE have to calculate the total number of pages
 		) );		
+	
+	}
+	function getCountDataState(){
+		global $wpdb ;
+		$sql = "select count(*) from ".$wpdb->prefix."states s ";
+		$cnt = $wpdb->get_var($sql);
+		if($cnt == null) $cnt = 0 ;
+		return $cnt ; 
 	
 	}
 	//Find Total Records
@@ -118,7 +150,9 @@ class States_Cities extends WP_List_Table {
 			switch($column_name){
 					case 'zip':
 					case 'state':
-					case 'stateCode':
+					case 'state_code':
+					case 'state_name':
+					case 'cities':
 					case 'city':
 							return $item[$column_name];
 					default:
@@ -138,11 +172,17 @@ class States_Cities extends WP_List_Table {
 							foreach($_GET['city'] as $val){
 								$this->bulkDelete($val);
 							}
-						}else{
+						}elseif($_GET['city'] != ""){
 							$this->bulkDelete($_GET['city']);
+						}elseif(is_array($_GET['state'])){
+							foreach($_GET['state'] as $val){
+								$this->bulkDeleteState($_GET['state']);
+							}
+						}elseif($_GET['state'] != ""){
+							$this->bulkDeleteState($_GET['state']);
 						}
-						
-            wp_die('Items deleted (or they would be if we had items to delete)!');
+						echo "<div class='updated'><p>Records have been deleted</p></div>";
+            //wp_die('Items deleted (or they would be if we had items to delete)!');
         }
 
         
@@ -152,10 +192,32 @@ class States_Cities extends WP_List_Table {
 		$sortable_columns = array(
 			'zip'  => array('zip',false),
 			'city' => array('city',false),
-			'stateCode'   => array('stateCode',false),
-			'state'   => array('state',false)
+			'state_code'   => array('state_code',false),
+			'state_name'   => array('state_name',false)
 		);
 		return $sortable_columns;
+	}
+	function get_sortable_columns_forState() {
+		$sortable_columns = array(
+			'state_code'   => array('state_code',false),
+			'state_name'   => array('state_name',false)
+		);
+		return $sortable_columns;
+	}
+	function column_state_name($item){
+		//Build row actions
+		$actions = array(
+				'edit'      => sprintf('<a href="?page=%s&action=%s&state=%s">Edit</a>',$_REQUEST['page'],'edit',$item['id']),
+				'delete'    => sprintf('<a href="?page=%s&action=%s&state=%s">Delete</a>',$_REQUEST['page'],'delete',$item['id']),
+		);
+		
+		//Return the title contents
+		return sprintf('%1$s <span style="color:silver">(id:%2$s)</span>%3$s',
+				/*$1%s*/ $item['state_name'],
+				/*$2%s*/ $item['id'],
+				/*$3%s*/ $this->row_actions($actions)
+		);
+	
 	}
     function column_city($item){
         
@@ -173,12 +235,19 @@ class States_Cities extends WP_List_Table {
         );
     }
     function column_cb($item){
+			
+			if($_REQUEST['page'] == "list_states")
+				$key = 'state' ;
+			else
+				$key = $this->_args['singular'] ;
+				
         return sprintf(
             '<input type="checkbox" name="%1$s[]" value="%2$s" />',
-            /*$1%s*/ $this->_args['singular'],  //Let's simply repurpose the table's singular label ("movie")
+            /*$1%s*/ $key ,  //Let's simply repurpose the table's singular label ("movie")
             /*$2%s*/ $item['id']                //The value of the checkbox should be the record's id
         );
     }    
+   
 	//This will be used for saving cities
 	public function saveCities($data){
 		global $wpdb;							
@@ -259,23 +328,59 @@ class States_Cities extends WP_List_Table {
 			$rows = array();
 			$current_page = $this->get_pagenum();
 			$offset = $this->per_page * ($current_page -1) ;
+			if(array_key_exists($_REQUEST['orderby'], $this->get_sortable_columns()) == true || 
+				$_REQUEST['orderby'] == ""){
+				$orderby = (!empty($_REQUEST['orderby'])) ? $_REQUEST['orderby'] : 'city'; //If no sort, default to title
+			}
+			if(in_array($_REQUEST['order'], array('','asc','desc')) == true){
+				$order = (!empty($_REQUEST['order'])) ? $_REQUEST['order'] : 'asc'; //If no order, default to asc
+			}
 			$sql = $wpdb->prepare("select c.id, c.zip,c.city,c.state_id, s.state_name from ".$wpdb->prefix."cities c , 
-																				".$wpdb->prefix."states s  
-																				where c.state_id=s.state_code order by state_name,city limit %d, %d" , 
-																				array( $offset ,  $this->per_page ));
+														".$wpdb->prefix."states s  
+														where c.state_id=s.state_code order by $orderby $order limit %d, %d" , 
+														$offset ,  $this->per_page );
 			$result = $wpdb->get_results ($sql );
 				foreach($result as $val){
 					$rows[] = array(
 					'id' => $val->id,
 					'zip' => $val->zip,
 					'city' =>$val->city,
-					'stateCode' => $val->state_id,
-					'state'=> $val->state_name ,
+					'state_code' => $val->state_id,
+					'state_name'=> $val->state_name ,
 					);
 				}
 				return $rows;	
 	}
-	
+
+	// Showing all the states and Cities
+	public function listDataStates(){
+			global $wpdb;
+			$rows = array();
+			$current_page = $this->get_pagenum();
+			$offset = $this->per_page * ($current_page -1) ;
+			if(array_key_exists($_REQUEST['orderby'], $this->get_sortable_columns()) == true || 
+				$_REQUEST['orderby'] == ""){
+				$orderby = (!empty($_REQUEST['orderby'])) ? $_REQUEST['orderby'] : 'state_name'; //If no sort, default to title
+			}
+			if(in_array($_REQUEST['order'], array('','asc','desc')) == true){
+				$order = (!empty($_REQUEST['order'])) ? $_REQUEST['order'] : 'asc'; //If no order, default to asc
+			}
+			$sql = $wpdb->prepare("select s.id, state_code, s.state_name, count(distinct c.id) as cities from ".$wpdb->prefix."states s  
+														left outer join ".$wpdb->prefix."cities c on c.state_id=s.state_code
+														group by state_code
+														order by $orderby $order limit %d, %d" , 
+														$offset ,  $this->per_page );
+			$result = $wpdb->get_results ($sql );
+				foreach($result as $val){
+					$rows[] = array(
+					'id' => $val->id,
+					'state_code' => $val->state_code,
+					'state_name'=> $val->state_name ,
+					'cities'=> $val->cities ,
+					);
+				}
+				return $rows;	
+	}	
 	//used for bluck delete
 	public function bulkDelete($id){
 		global $wpdb;
@@ -285,10 +390,33 @@ class States_Cities extends WP_List_Table {
 				array('%d')
 			);	
 	}
+	public function bulkDeleteState($id){
+		global $wpdb;
+		$sql = "select state_code from ".$wpdb->prefix."states s ";
+		$cnt = $wpdb->get_var($sql);
+		if($cnt == null) return 0  ;	
+		
+		$wpdb->delete(
+			"{$wpdb->prefix}cities",
+			array('state_id' => $cnt ),
+			array('%d')
+		);	
+		$wpdb->delete(
+			"{$wpdb->prefix}states",
+			array('id' => $id ),
+			array('%d')
+		);			
+		
+	}
   public function showListStates(){
-		//$ob1 = new States_Cities() ;
-		$this->prepare_items(); 
+		echo '<h2>List of States</h2>';
+		echo '<form id="movies-filter" method="get">'; ?>
+            <!-- For plugins, we also need to ensure that the form posts back to our current page -->
+            <input type="hidden" name="page" value="<?php echo $_REQUEST['page'] ?>" />		
+    <?php
+		$this->prepare_items_forState(); 
 		$this->display();
+		echo '</form>';
   }	
   public function showListCity(){
 		echo '<h2>List of Cities</h2>';
@@ -336,27 +464,17 @@ class States_Cities extends WP_List_Table {
 	
 	//Form Menu
 	function my_menu_pages(){
-		$import_state_city	=	add_menu_page('States Cities', 'States Cities', 'manage_options', 
-		'states_cities', array(&$this, 'init') );
+		
+		add_menu_page('states_cities', 'Import Cities', 'manage_options', 'states_cities', array(&$this, 'init') );
 		add_submenu_page('states_cities', 'List of Cities', 'List of Cities', 'manage_options', 'list_cities', array( new States_Cities(), 'showListCity'));
-		//add_submenu_page('states_cities', 'List of States', 'List of States', 'manage_options', 'list_states', array( new States_Cities(),'showListState'));
+		add_submenu_page('states_cities', 'List of States', 'List of States', 'manage_options', 'list_states', array( new States_Cities(),'showListStates'));
 	}
 
-	
-		//Create menu
-		function menu_list(){
 
-		}
 	} // Class
 	
-	add_action( 'importcsv', 'parseCSV' );	
 	$ob = new States_Cities ;
-	
-	function parseCSV(){
-		echo "Manish ";
-		die;
-		$ob->parseCSV();
-	}
+
 }// If
